Add support for ed25519 keys.
This commit adds the ed25519 implementation as a module to prevent
conflicts with similarly named objects from the p384 crate. Additionally
ed25519 only supports sha512 hashes when generating signatures so we
update the `signature_algorithm` function from `KeyPair` to return a
`Result`. This allows us to return an error if the config specifies an
incompatible digest algorithm.