Add support for CSR generation.
PKCS#10 / RFC 2986 certificate signing requests (CSR) are made up of:
- a version number
- an x501 distinguished name (the subject)
- a subject public key info (SKPI) structure
- an optional collection of attributes

CSRs are a subset of the x509 certificate structure we already generate
and so this commit draws heavily from the existing code. Some
refactoring was required to prevent duplication of the code that loads
the required keys and config entities. The reset of this commit is a
pretty mechanical conversion of config entries from the KDL file, to the
rust types from spki & x509-cert crates.