rsa: Limit RSA key generation to 2 primes.
In versions of the rsa crate >= 0.9 the
`generate_multi_prime_key_with_exp` functions are no longer public.
Their reasons for doing so are here:
https://github.com/RustCrypto/RSA/pull/304 but the TLDR; is: so no one
hurts themselves.

The closest analog available is `RsaPrivateKey::new_with_exp` but it
doesn't allow us to specify the number of primes. Internally it hard
codes this value to '2' which is our current default. None of our
example configs set this value and they probably shouldn't so this
commit removes the `num-primes` config option from the RsaKeyConfig.

This changes the config file structure so this commit bumps the version
number too.