Upgrade RSA crate to 0.9.3
Upstream has changed the way they use the zeroize crate as described
here: https://github.com/RustCrypto/RSA/issues/285. Because of these
changes we don't need to (and can't) wrap the RsaPrivateKey type in the
`Zeroizing` wrapper. By no longer using this wraper we don't need to
dereference it before passing it around.