Add support for the DiceTcbInfo x509 extension.
This extension is defined in the TCG Dice Attestation Architecture spec
v1.0.0 in §6.1.1. We only support the `fwids` data field from the
DiceTcbInfo extension. Additional fields may be implemented as needed.

The TCG spec states that certs with the DiceTcbInfo extension MUST
include the RFC 5288 AuthorityKeyIdentifier. We do not enforce this
restriction but our implementation does not preclude the generation of a
conformant cert chain from the KDL.