Add KDL node for `CertificateList`s & command to generate them
It's common for lists of certificates to be concatenated and
stored or transferred as a single file / unit. Different applications /
protocols may expect or produce a particular ordering. In rare cases
certs in a list may be out of order, or even contain multiple branches.

To support as many use-cases as possible we do not enforce any
particular ordering or relationship between the certificates in a
certificate list. We simply concatenate them in the order specified.

A `CertificateList` is defined in KDL as:
```kdl
certificate-list "name" \
    "cert-name-1" \
    "cert-name-2"
```

The output file name prefix is taken from the first attribute of the KDL
("name" in this example). The suffix `.certlist.pem` is appended to this
`name`. The remaining attributes are the string names assigned to
`certificate` nodes defined elsewhere in the document.