Add KDL node for `CertificateList`s & command to generate them
It's common for ordered lists of certificates to be concatenated and
stored or transferred as a single file / unit. Different applications /
protocols may expect a particular ordering. To support as many use-cases
as possible we do not enforce any particular ordering or relationship
between the certificates in a certificate list. We simply concatenate
them in the order specified.

So far we only encode files as PEM and so we use the file extension
`.certlist.pem`.

A `CertificateList` is defined in the KDL as:
```kdl
certificate-list "name" \
    "cert-name-1" \
    "cert-name-2"
```
The output file name prefix is taken from the first attribute of the KDL
("name" in this example). The remaining attributes are the string names
assigned `certificate` nodes defined elsewhere in the document.