wip: add top level node for PkiPath, construct it and write to a file
RFC 6066 defines pkipaths as DER encoded and identified by the
`application/pkix-pkipath` MIME type. When written to a file the
extension `.pkipath` per:
https://www.iana.org/assignments/media-types/application/pkix-pkipath
So far we only encode files as PEM and so we use the file extension
`.pkipath.pem`.

A PkiPath is defined in the KDL as:
```kdl
pki-path "name" {
    cert "cert1-name"
    cert "cert2-name"
}
```

I'd prefer to use something like
```kdl
pki-path "name" {
    "cert1-name"
    "cert2-name"
}
```

but I haven't figured out how to use the knuffel derives to achieve this
yet. It's trivial to use attributes instead of children as a way to get
rid of the child `cert` nodes but it's not particularly readable:
```kdl
pki-path name="foo" "cert1-name" "cert2-name"
```
but I don't think that's particularly readable.